Add files via upload
The above file is a ping pong game made using turtle module and library. The game starts with a triangle ball and the there can be two players to play the game. Namely, they can be lefty and righty. The score limit is one, so if the ball is saved by one player the game is over and the other player wins the game.